+++ api-gateway-internal/gateway/src/main/resources/application.yaml
@@ -11,3 +11,37 @@ spring:
               uri: http://localhost:8081
               predicates:
                 - Path=/trails_api/**
+            - id: events-service
+              uri: http://provider-events:8083
+              predicates:
+                - Path=/events-api/**
+              filters:
+                - RewritePath=/events-api/(?<segment>.*), /$\{segment}
+                
+            - id: food-service
+              uri: http://provider-food:8084
+              predicates:
+                - Path=/food-api/**
+              filters:
+                - RewritePath=/food-api/(?<segment>.*), /$\{segment}
+                
+            - id: trails-service
+              uri: http://provider-trails:8085
+              predicates:
+                - Path=/trails-api/**
+              filters:
+                - RewritePath=/trails-api/(?<segment>.*), /$\{segment}
+                
+            - id: itinerary-aggregator
+              uri: http://itinerary-aggregator:8082
+              predicates:
+                - Path=/itinerary-api/**
+              filters:
+                - RewritePath=/itinerary-api/(?<segment>.*), /$\{segment}
+                
+            - id: availability-coordinator
+              uri: http://availability-coordinator:8081
+              predicates:
+                - Path=/availability-api/**
+              filters:
+                - RewritePath=/availability-api/(?<segment>.*), /$\{segment}
