--- provider-food/resources/data.sql
@@ -1,6 +1,0 @@
-INSERT INTO restaurant(id, name, location, rating, specialties, cuisine_type) 
-VALUES
-('r1', 'Taverna Ducale', 'Sulmona', 4.7, 'Arrosticini,Maccheroni alla chitarra', 'TRADITIONAL'),
-('r2', 'Ristorante del Lago', 'Scanno', 4.5, 'Trota,Coregone', 'SEAFOOD'),
-('r3', 'La Grotta dei Sapori', 'Roccaraso', 4.8, 'Polenta,Cinghiale', 'MOUNTAIN'),
-('r4', 'Il Calesse', 'Pescocostanzo', 4.6, 'Agnello,Formaggi locali', 'TRADITIONAL');
