+++ provider-food/src/main/java/com/example/providerfood/model/LocalDish.java
@@ -1,5 +1,4 @@
 package com.example.providerfood.model;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
