+++ provider-food/src/main/resources/application.yml
@@ -16,8 +16,11 @@ spring:
   jpa:
     database-platform: org.hibernate.dialect.H2Dialect
     hibernate:
-      ddl-auto: create-drop
+      ddl-auto: create
     show-sql: true
+  sql:
+    init:
+      mode: never
 
 springdoc:
   swagger-ui:
