+++ provider-food/src/main/resources/data.sql
@@ -1,0 +1,6 @@
+INSERT INTO restaurant(id, name, location, rating, cuisine_type, specialties) 
+VALUES
+('r1', 'Taverna Ducale', 'Sulmona', 4.7, 'TRADITIONAL', 'Arrosticini,Maccheroni alla chitarra'),
+('r2', 'Ristorante del Lago', 'Scanno', 4.5, 'SEAFOOD', 'Trota,Coregone'),
+('r3', 'La Grotta dei Sapori', 'Roccaraso', 4.8, 'MOUNTAIN', 'Polenta,Cinghiale'),
+('r4', 'Il Calesse', 'Pescocostanzo', 4.6, 'TRADITIONAL', 'Agnello,Formaggi locali');
